+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_validators.py
@@ -422,41 +422,16 @@ def _validate_vm_create_storage_profile(cmd, namespace, for_scale_set=False):
     if not namespace.os_type:
         namespace.os_type = 'windows' if 'windows' in namespace.os_offer.lower() else 'linux'
 
-    namespace.os_caching = _process_disk_caching_value(cmd, namespace.os_caching, is_os_disk=True)
-    namespace.data_caching = _process_disk_caching_value(cmd, namespace.data_caching, is_os_disk=False)
-
-
-def _process_disk_caching_value(cmd, disk_caching, is_os_disk):
-    from azure.cli.core.profiles import ResourceType
-    CachingTypes = cmd.get_models('CachingTypes', resource_type=ResourceType.MGMT_COMPUTE)
-
-    if is_os_disk:
-        if not disk_caching:
-            return CachingTypes.read_write.value
-        # value check is done at arg-parse layer through enum choice list, so skipping here
-        return disk_caching
-
-    if disk_caching:
-        disk_caching_values = [x.value for x in CachingTypes]
-        disk_caching_values_lower = [x.value.lower() for x in CachingTypes]
-        invalid_caching_val_err = 'usage error: please use data disk caching value from ' + '|'.join(
-            disk_caching_values)
-        using_luns = [x for x in disk_caching if '=' in x]
-        if len(disk_caching) > 1:
-            # verify if we have 2+ entries, all should have the format of '<lun>=<value>'
-            if len(using_luns) != len(disk_caching):
-                raise CLIError('usage error: --data-disk-caching VALUE | --data-disk-caching LUN=VALUE LUN2=VALUE2 ...')
-        for x in disk_caching:
-            c = x
-            if '=' in x:
-                lun, c = x.split('=', 1)
-                try:
-                    lun = int(lun)
-                except ValueError:
-                    raise CLIError("usage error: LUN used in --data-disk-caching must be an integer")
-            if c.lower() not in disk_caching_values_lower:
-                raise CLIError(invalid_caching_val_err.format('data'))
-    return disk_caching
+    from ._vm_utils import normalize_disk_info
+    # accelnet and attach_data_disks are not exposed yet for VMSS, so use 'getattr' to avoid crash
+    write_accelerator_settings = getattr(namespace, 'write_accelerator', None)
+    namespace.disk_info = normalize_disk_info(image_data_disks=namespace.image_data_disks,
+                                              data_disk_sizes_gb=namespace.data_disk_sizes_gb,
+                                              attach_data_disks=getattr(namespace, 'attach_data_disks', []),
+                                              storage_sku=namespace.storage_sku,
+                                              os_disk_caching=namespace.os_caching,
+                                              data_disk_cachings=namespace.data_caching,
+                                              write_accelerator_settings=write_accelerator_settings)
 
 
 def _validate_vm_create_storage_account(cmd, namespace):