+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -4,6 +4,7 @@ Release History
 ===============
 2.0.31
 ++++++
+* vm: support write accelerator
 * vmss: expose `az vmss perform-maintenance`
 * `vm diagnostics set`: detect VM's OS type reliably
 * `vm resize`: check if the requested size is different than currently set and update only on change